:fire: Remove ValueType
